requirements.txt
altair
bs4
ipykernel
nltk
pylint
pytest
python-dotenv
pandas
requests
spacy